+++ app/helpers/observation_matrix_columns_helper.rb
@@ -5,6 +5,11 @@ module ObservationMatrixColumnsHelper
     matrix_column.descriptor.name
   end
 
+  def label_for_observation_matrix_column(matrix_column)
+    return nil if matrix_column.nil?
+    matrix_column.descriptor.name
+  end
+
   def observation_matrix_column_link(matrix_column)
     return nil if matrix_column.nil?
     link_to(observation_matrix_column_tag(matrix_column).html_safe, matrix_column)