+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 - Rake task to generate docs.taxonworks.org Data documentation [#2352]
 - Confirmation window for delete documentation in radial annotator [#2820]
 - Drag and drop to sort predicates in project preferences [#2821]
+- Endpoints for observation matrix row and column labels [#2800]
 
 ### Changed
 - Browse nomenclature task was renamed to Browse nomenclature and classifications [#2638]
@@ -33,6 +34,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 - Download link doesn't work in data list view
 - Edit in Browse collecting event [#2814]
 
+[#2800]: https://github.com/SpeciesFileGroup/taxonworks/issues/2800
 [#2352]: https://github.com/SpeciesFileGroup/taxonworks/issues/2552
 [#2550]: https://github.com/SpeciesFileGroup/taxonworks/issues/2550
 [#2790]: https://github.com/SpeciesFileGroup/taxonworks/issues/2790